wazevo(ssa): refactors pass execution
